requirements.txt
langchain
streamlit
google-generativeai
python-dotenv
PyPDF2
chromadb
langchain_google_genai
